requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beanie==1.7.2
chardet==4.0.0
click==7.1.2
discord.py @ git+https://github.com/iDevision/enhanced-discord.py.git@e621d5288eeec2472f248efe8ce5581348dadf8a
dnspython==2.1.0
idna==3.3
motor==2.5.1
multidict==5.2.0
pydantic==1.8.2
pymongo==3.12.1
python-Levenshtein==0.12.2
thefuzz==0.19.0
toml==0.10.2
typing_extensions==4.0.0
yarl==1.7.2
